+++ helixdb/Cargo.toml
@@ -23,7 +23,6 @@ chrono = "0.4.39"
 flume = "0.11.1"
 rayon = "1.8.0"
 itertools = "0.14.0"
-colored = "2.1.0"
 get_routes = { version = "0.1.0", path = "../get_routes" }
 
 # Compiler
