+++ helixdb/src/helixc/analyzer/analyzer.rs
@@ -1,7 +1,5 @@
 //! Semantic analyzer for Helix‑QL.
-
-use colored::Colorize;
-
+use super::{fix::Fix, pretty};
 use crate::{
     helix_engine::graph_core::ops::source::add_e::EdgeType,
     helixc::{
@@ -9,12 +7,12 @@ use crate::{
             bool_op::{BoolOp, Eq, Gt, Gte, Lt, Lte, Neq},
             generator_types::{
                 Assignment as GeneratedAssignment, BoExp, Drop as GeneratedDrop,
-                ForEach as GeneratedForEach, ForLoopInVariable, ForVariable, IdentifierType,
+                ForEach as GeneratedForEach, ForLoopInVariable, ForVariable,
                 Parameter as GeneratedParameter, Query as GeneratedQuery, ReturnValue,
                 ReturnValueExpr, Source as GeneratedSource, Statement as GeneratedStatement,
             },
             object_remapping_generation::{
-                ExcludeField, FieldRemapping, IdentifierRemapping, ObjectRemapping, Remapping,
+                ExcludeField, IdentifierRemapping, ObjectRemapping, Remapping,
                 RemappingType, TraversalRemapping, ValueRemapping,
             },
             source_steps::{
@@ -28,23 +26,18 @@ use crate::{
                 WhereExists, WhereRef,
             },
             utils::{
-                GenRef, GeneratedType, GeneratedValue, RustType as GeneratedRustType, Separator,
+                GenRef, GeneratedValue, Separator,
             },
         },
         parser::{
-            helix_parser::{ShortestPath, *},
+            helix_parser::*,
             location::Loc,
         },
     },
     protocol::{date::Date, value::Value},
+    utils::styled_string::StyledString,
 };
-
-use std::{
-    collections::{HashMap, HashSet},
-    ops::{ControlFlow, Deref},
-};
-
-use super::{fix::Fix, pretty};
+use std::collections::{HashMap, HashSet};
 
 /// A single diagnostic to be surfaced to the editor.
 #[derive(Debug, Clone)]
@@ -1714,10 +1707,10 @@ impl<'a> Ctx<'a> {
                         self.push_query_err(
                             q,
                             ex.loc.clone(),
-                            "exclude is only valid as the last step in a traversal, 
+                            "exclude is only valid as the last step in a traversal,
                             or as the step before an object remapping or closure"
                                 .to_string(),
-                            "move exclude steps to the end of the traversal, 
+                            "move exclude steps to the end of the traversal,
                             or remove the traversal steps following the exclude"
                                 .to_string(),
                         );
@@ -4352,7 +4345,7 @@ impl<'a> Ctx<'a> {
         println!("{:?}", name);
         if self.is_param(q, name) {
             println!("{:?}", name);
-            GeneratedValue::Parameter(GenRef::Ref(format!("data.{}", name)))
+            GeneratedValue::Parameter(GenRef::Std(format!("data.{}", name)))
         } else {
             GeneratedValue::Identifier(GenRef::Std(name.to_string()))
         }
@@ -4413,7 +4406,7 @@ impl Type {
     }
 
     /// Same, but returns an owned clone for convenience.
-    fn cloned_base(&self) -> Type {
+    fn cloned_base(&self) -> Type { // TODO: never used?
         match self {
             Type::Anonymous(inner) => inner.cloned_base(),
             _ => self.clone(),
@@ -4432,333 +4425,3 @@ impl<'a> From<&'a FieldType> for Type {
     }
 }
 
-// ---------------------------------
-// Tests
-// ---------------------------------
-#[cfg(test)]
-mod analyzer_tests {
-    use super::*;
-    use crate::helixc::parser::helix_parser::HelixParser;
-
-    /// Convenience helper – parse text and return diagnostics.
-    fn run(src: &str) -> Vec<Diagnostic> {
-        let input = write_to_temp_file(vec![src]);
-        let parsed = HelixParser::parse_source(&input)
-            .expect("parser should succeed – these tests are for the analyzer");
-        analyze(&parsed).0
-    }
-
-    #[test]
-    fn reports_unknown_node_in_edge() {
-        let hx = r#"
-            E::Likes {
-                From: User,
-                To: Post,
-                Properties: {}
-            }
-        "#;
-        let diags = run(hx);
-        assert!(
-            diags
-                .iter()
-                .any(|d| d.message.contains("not a declared node type")),
-            "expected a diagnostic about undeclared node types, got: {:?}",
-            diags
-        );
-    }
-
-    #[test]
-    fn detects_redeclared_variable() {
-        let hx = r#"
-            N::User { Name: String }
-
-            QUERY dupVar() =>
-                u <- N<User>
-                u <- N<User>
-                RETURN u
-        "#;
-        let diags = run(hx);
-        assert!(
-            diags.iter().any(|d| d.message.contains("already declared")),
-            "expected a diagnostic about variable redeclaration, got: {:?}",
-            diags
-        );
-    }
-
-    #[test]
-    fn flags_invalid_property_access() {
-        let hx = r#"
-            N::User { name: String }
-
-            QUERY badField() =>
-                u <- N<User>
-                n <- u::{age}
-                RETURN n
-        "#;
-        let diags = run(hx);
-        assert!(
-            diags
-                .iter()
-                .any(|d| d.message.contains("is not a field of node")),
-            "expected a diagnostic about invalid property access, got: {:?}",
-            diags
-        );
-    }
-
-    #[test]
-    fn traversal_step_order_enforced() {
-        let hx = r#"
-            N::User { name: String }
-
-            QUERY wrongStep() =>
-                e <- N<User>::FromN   // OutN on nodes is illegal (needs Edge)
-                RETURN e
-        "#;
-        let diags = run(hx);
-        assert!(
-            diags
-                .iter()
-                .any(|d| d.message.contains("cannot follow a step that returns")),
-            "expected a diagnostic about illegal step ordering, got: {:?}",
-            diags
-        );
-    }
-
-    #[test]
-    fn clean_query_produces_no_diagnostics() {
-        let hx = r#"
-            N::User { name: String }
-            N::Post { title: String }
-            E::Wrote {
-                From: User,
-                To: Post,
-                Properties: {}
-            }
-
-            QUERY ok(hey: User) =>
-                u <- N<User>
-                p <- u::Out<Wrote>
-                RETURN p::!{title}::{title}
-        "#;
-        let diags = run(hx);
-        for d in diags.iter() {
-            println!("{}", d.render(hx, "query.hx"));
-        }
-        assert!(
-            diags.is_empty(),
-            "expected no diagnostics, got: {:?}",
-            diags
-        );
-    }
-
-    #[test]
-    fn validates_edge_properties() {
-        let hx = r#"
-            N::User { name: String }
-            N::Post { title: String }
-            E::Wrote {
-                From: User,
-                To: Post,
-                Properties: {
-                    date: String,
-                    likes: I32
-                }
-            }
-
-            QUERY badEdgeField() =>
-                e <- N<User>::OutE<Wrote>
-                n <- e::{invalid_field}
-                RETURN n
-        "#;
-        let diags = run(hx);
-        for d in diags.iter() {
-            println!("{}", d.render(hx, "query.hx"));
-        }
-        assert!(
-            diags
-                .iter()
-                .any(|d| d.message.contains("is not a field of edge")),
-            "expected a diagnostic about invalid edge property access, got: {:?}",
-            diags
-        );
-    }
-
-    #[test]
-    fn validates_node_properties() {
-        let hx = r#"
-            N::User { name: String }
-            N::Post { title: String }
-
-            QUERY badNodeField() =>
-                n <- N<User>::{invalid_field}
-                RETURN n
-        "#;
-        let diags = run(hx);
-        for d in diags.iter() {
-            println!("{}", d.render(hx, "query.hx"));
-        }
-        assert!(
-            diags
-                .iter()
-                .any(|d| d.message.contains("is not a field of edge")),
-            "expected a diagnostic about invalid edge property access, got: {:?}",
-            diags
-        );
-    }
-
-    #[test]
-    fn validates_vector_properties() {
-        let hx = r#"
-            V::UserEmbedding {
-                content: String
-            }
-            
-            QUERY badVectorField(vec: [F64], content: String) =>
-                v <- AddV<UserEmbedding>(vec,{content: content})
-                RETURN v
-        "#;
-        let diags = run(hx);
-        for d in diags.iter() {
-            println!("{}", d.render(hx, "query.hx"));
-        }
-        assert!(
-            diags
-                .iter()
-                .any(|d| d.message.contains("is not a field of vector")),
-            "expected a diagnostic about invalid vector property access, got: {:?}",
-            diags
-        );
-    }
-
-    #[test]
-    fn handles_untyped_nodes() {
-        let hx = r#"
-            N::User { name: String }
-
-            QUERY untypedNode() =>
-                u <- N<User>::{some_field}
-                RETURN n
-        "#;
-        let diags = run(hx);
-        for d in diags.iter() {
-            println!("{}", d.render(hx, "query.hx"));
-        }
-        assert!(
-            diags.is_empty(),
-            "expected no diagnostics for untyped node access, got: {:?}",
-            diags
-        );
-    }
-
-    #[test]
-    fn respects_excluded_fields() {
-        let hx = r#"
-            N::User { name: String, age: I32 }
-
-            QUERY excludedField() =>
-                u <- N<User>
-                n <- u::!{name}::{name}
-                RETURN n
-        "#;
-        let diags = run(hx);
-        for d in diags.iter() {
-            println!("{}", d.render(hx, "query.hx"));
-        }
-        assert!(
-            diags
-                .iter()
-                .any(|d| d.message.contains("was previously excluded")),
-            "expected a diagnostic about accessing excluded field, got: {:?}",
-            diags
-        );
-    }
-
-    #[test]
-    fn validates_add_node_fields() {
-        let hx = r#"
-            N::User { name: String }
-
-            QUERY badAddNodeField() =>
-                n <- AddN<User>({invalid_field: "test"})
-                RETURN n
-        "#;
-        let diags = run(hx);
-        assert!(
-            diags
-                .iter()
-                .any(|d| d.message.contains("is not a field of node")),
-            "expected a diagnostic about invalid node field, got: {:?}",
-            diags
-        );
-    }
-
-    #[test]
-    fn validates_add_edge_fields() {
-        let hx = r#"
-            N::User { name: String }
-            N::Post { title: String }
-            E::Wrote {
-                From: User,
-                To: Post,
-                Properties: {
-                    date: String
-                }
-            }
-
-            QUERY badAddEdgeField() =>
-                n1 <- AddN<User>({name: "test"})
-                n2 <- AddN<Post>({title: "test"})
-                e <- AddE<Wrote>({invalid_field: "test"})::To(n1)::From(n2)
-                RETURN e
-        "#;
-        let diags = run(hx);
-        assert!(
-            diags
-                .iter()
-                .any(|d| d.message.contains("is not a field of edge")),
-            "expected a diagnostic about invalid edge field, got: {:?}",
-            diags
-        );
-    }
-
-    #[test]
-    fn validates_add_vector_fields() {
-        let hx = r#"
-            V::UserEmbedding {
-                content: String
-            }
-            
-            QUERY badAddVectorField() =>
-                v <- AddV<UserEmbedding>([1.0, 2.0], {invalid_field: "test"})
-                RETURN v
-        "#;
-        let diags = run(hx);
-        assert!(
-            diags
-                .iter()
-                .any(|d| d.message.contains("is not a valid vector field")),
-            "expected a diagnostic about invalid vector field, got: {:?}",
-            diags
-        );
-    }
-
-    #[test]
-    fn validate_boolean_comparison() {
-        let hx = r#"
-            N::User { name: String }
-
-            QUERY booleanComparison() =>
-                a <- N<User>::WHERE(_::{name}::EQ(10))
-                RETURN a
-        "#;
-        let diags = run(hx);
-        for d in diags.iter() {
-            println!("{}", d.render(hx, "query.hx"));
-        }
-        assert!(
-            diags.is_empty(),
-            "expected no diagnostics, got: {:?}",
-            diags
-        );
-    }
-}
