+++ helixdb/src/helixc/analyzer/mod.rs
@@ -2,3 +2,7 @@ pub mod analyzer;
 pub mod pretty;
 pub mod fix;
 pub mod types;
+
+#[cfg(test)]
+mod analyzer_tests;
+
