+++ helixdb/src/helixc/generator/mod.rs
@@ -5,3 +5,7 @@ pub mod source_steps;
 pub mod traversal_steps;
 pub mod tsdisplay;
 pub mod utils;
+
+#[cfg(test)]
+mod generator_tests;
+
