+++ helixdb/src/helixc/generator/utils.rs
@@ -329,7 +329,7 @@ use helixdb::{
         },
         vectors::{insert::InsertVAdapter, search::SearchVAdapter, brute_force_search::BruteForceSearchVAdapter},
         bm25::search_bm25::SearchBM25Adapter,
-        
+
     },
     helix_engine::types::GraphError,
     helix_gateway::router::router::HandlerInput,
