+++ helixdb/src/helixc/parser/mod.rs
@@ -1,3 +1,7 @@
 pub mod helix_parser;
 pub mod location;
 pub mod parser_methods;
+
+#[cfg(test)]
+mod parser_tests;
+
