+++ helixdb/src/lib.rs
@@ -5,3 +5,5 @@ pub mod helixc;
 #[cfg(feature = "ingestion")]
 pub mod ingestion_engine;
 pub mod protocol;
+pub mod utils;
+
