+++ helixdb/src/utils/mod.rs
@@ -1,0 +1,2 @@
+pub mod styled_string;
+
